Refactored from static classes. Renaming: - AuthSession => AuthClient - Authenticator => AuthSession
